demo/boardgames/requirements.txt
appdirs==1.4.3
Django==1.8
django-crispy-forms==1.6.1
django-debug-toolbar==1.7
packaging==16.8
pyparsing==2.2.0
six==1.10.0
sqlparse==0.2.3
